Differences from the first version:
Working with the database is implemented through Spring Data JPA.
All data output is divided into pages.
There is a possibility to search for books in the database.
Books are issued in sorted order (if the corresponding key is specified in the URL)

The project is published as a separate one, since the work with the database is carried out using a different technology